verse ject tope

+++ stlx.h
@@ -38,7 +38,7 @@ struct SafeState {
         if (pthread_mutex_lock(&mutex) != 0) {std::cerr << "cannot lock mutex!" << std::endl; exit(-1);}
         struct timespec ts;
         if (clock_gettime(CLOCK_REALTIME, &ts) == -1) {std::cerr << "cannot get time!" << std::endl; exit(-1);}
-        dif += ts.tv_sec + ts.tv_nsec/NANOSECONDS;
+        dif += ts.tv_sec + (double)ts.tv_nsec/NANOSECONDS;
         double tsi, tsf; tsf = modf(dif, &tsi); ts.tv_sec = tsi; ts.tv_nsec = tsf*NANOSECONDS;
         int val = 0; while (count == 0 && val == 0) if ((val = pthread_cond_timedwait(&condit,&mutex,&ts)) != 0 && val != ETIMEDOUT) {std::cerr << "cannot timed wait!" << std::endl; exit(-1);}
         if (val != ETIMEDOUT && count > 0) count -= 1;
